refactor: move classes to src subfolder
